Works
This is a difficult one!  `inWidget` is recursive and defined in terms
of `r` which is itself recursive.  Furthermore, the main block of
effectful code is then passed to `Internal.unsafeUnEff` so we lose the
ability to infer its type.  Adding the @es type parameter helps
inference.  For some reason, adding a partial type signature for `r`
helps too.

Bluefin doesn't do well here.  I'm not sure I can explain that in
terms other than this is a rather strange function!